Extract Config to separate file
